Drawing

Types

data Drawing

type Point = (Double, Double)

A point on the x-y plane.

data Color

type TextAnchor = Text

Basic drawings

blank :: Drawing

A blank drawing, with nothing in it.

polyline :: forall f. Foldable f => f Point -> Drawing

A sequence of line segments, with given endpoints.

polygon :: forall f. Foldable f => f Point -> Drawing

A thin polygon with these points as vertices.

solidPolygon :: forall f. Foldable f => f Point -> Drawing

A solid polygon with these points as vertices.

rectangle :: Double -> Double -> Drawing

solidRectangle :: Double -> Double -> Drawing

circle :: Double -> Drawing

solidCircle :: Double -> Drawing

text :: Text -> Drawing

text s = atext middleAnchor s

atext :: TextAnchor -> Text -> Drawing

coordinatePlane :: Drawing

A coordinate plane. Adding this to your pictures can help you measure distances more accurately.

Example:

main = putSvg (coordinatePlane <> myPicture)
myPicture = ...

Modification of drawings

colored :: Color -> Drawing -> Drawing

A drawing drawn with this color.

translated :: Double -> Double -> Drawing -> Drawing

A drawing translated by the given x and y coordinates.

translated dx dy = transform (translation (dx, dy))

scaled :: Double -> Double -> Drawing -> Drawing

A drawing scaled by the given x and y factors. Scaling by a negative factor also reflects across that axis.

scaled sx sy = transform (scaling sx sy)

dilated :: Double -> Drawing -> Drawing

A drawing scaled uniformly by the given factor. Dilating by a negative factor also reflects across the origin.

dilated s = scaled s s

rotated :: Double -> Drawing -> Drawing

A drawing rotated by this angle about the origin.

Angles are in radians.

rotated a = transform (rotation a)

xSkewed :: Double -> Drawing -> Drawing

xSkewed a = transform (skewX a)

ySkewed :: Double -> Drawing -> Drawing

ySkewed a = transform (skewY a)

Colors

white :: Color

black :: Color

gray :: Color

red :: Color

green :: Color

lime :: Color

blue :: Color

yellow :: Color

orange :: Color

brown :: Color

fromRGB :: Double -> Double -> Double -> Color

fromHSL :: Double -> Double -> Double -> Color

toRGB :: Color -> (Double, Double, Double)

toHSL :: Color -> (Double, Double, Double)

hue :: Color -> Double

saturation :: Color -> Double

luminosity :: Color -> Double

lighter :: Double -> Color -> Color

light :: Color -> Color

dark :: Color -> Color

brighter :: Double -> Color -> Color

bright :: Color -> Color

dull :: Color -> Color

Text anchors

startAnchor :: TextAnchor

middleAnchor :: TextAnchor

endAnchor :: TextAnchor

Composition

(<>) :: Semigroup a => a -> a -> a infixr 6

An associative operation.

>>> [1,2,3] <> [4,5,6]
[1,2,3,4,5,6]

Affine transformations

data Transform

An affine transformation.

An affine transformation is any transformation that preserves collinearity (i.e., all points lying on a line initially still lie on a line after transformation) and ratios of distances (e.g., the midpoint of a line segment remains the midpoint after transformation). An affine transformation is also called an affinity.

translation :: (Double, Double) -> Transform

A translation by the given distance vector.

scaling :: Double -> Double -> Transform

A scaling by the given x and y factors. Scaling by a negative factor also reflects across that axis.

dilation :: Double -> Transform

An uniformly scaling by the given factor. Dilating by a negative factor also reflects across the origin.

dilated s = scaled s s

rotation :: Double -> Transform

A rotation by the given angle about the origin. Angles are in radians.

skewX :: Double -> Transform

skewY :: Double -> Transform

transform :: Transform -> Drawing -> Drawing

Apply an affine transformation to a drawing.

transformPoint :: Transform -> Point -> Point

Apply an affine transformation to a point.

transformVector :: Transform -> (Double, Double) -> (Double, Double)

Apply the linear part of an affine transformation (that's a linear tranformation, ignoring the translation part) to a vector.

I/O and interaction

module Drawing.IO

drawingOf

Arguments

:: Double
The width (in drawing units) of the view.

-> Double
The height (in drawing units) of the view.

-> Drawing

-> IO ()

A simple drawing CGI program.

animationOf

Arguments

:: Double
The width (in drawing units) of the view.

-> Double
The height (in drawing units) of the view.

-> (Time -> Drawing)
Function to produce the next frame of animation. It is passed the time in seconds since the program started.

-> IO ()

A CGI program which shows an animation, with a picture for each time given by the parameter.

activityOf

Arguments

:: Double
The width (in drawing units) of the view.

-> Double
The height (in drawing units) of the view.

-> a
The initial state of the activity.

-> (Event -> a -> a)
The event handling function, which updates the state given an event.

-> (a -> Drawing)
The visualization function, which converts the state into a picture to display.

-> IO ()

An interactive CGI program that responds to Events. Activities can interact with the user, change over time, and remember information about the past.

splitTime :: Time -> (Time -> a -> a) -> Time -> a -> a

data Event

Constructors

KeyDown !Text

KeyUp !Text

MouseDown !Point

MouseUp !Point

MouseMove !Point

TimePassing !Time
The argument is the time in seconds since the previous TimePassing event.

type Time = Double